+++ src/main/java/br/com/zalf/prolog/webservice/integracao/sistema/SistemasFactory.java
@@ -3,6 +3,9 @@ package br.com.zalf.prolog.webservice.integracao.sistema;
 import br.com.zalf.prolog.webservice.integracao.IntegradorProLog;
 import br.com.zalf.prolog.webservice.integracao.avacorpavilan.AvaCorpAvilan;
 import br.com.zalf.prolog.webservice.integracao.avacorpavilan.requester.AvaCorpAvilanRequesterImpl;
+import br.com.zalf.prolog.webservice.integracao.protheusrodalog.ProtheusRodalogRequesterImpl;
+import br.com.zalf.prolog.webservice.integracao.protheusrodalog.SistemaProtheusRodalog;
+import br.com.zalf.prolog.webservice.integracao.transport.SistemaTransportTranslecchi;
 import com.sun.istack.internal.NotNull;
 
 /**
@@ -22,6 +25,17 @@ public final class SistemasFactory {
         switch (sistemaKey) {
             case AVACORP_AVILAN:
                 return new AvaCorpAvilan(new AvaCorpAvilanRequesterImpl(), sistemaKey, integradorProLog, userToken);
+            case TRANSPORT_TRANSLECCHI:
+                return new SistemaTransportTranslecchi(
+                        sistemaKey,
+                        integradorProLog,
+                        userToken);
+            case PROTHEUS_RODALOG:
+                return new SistemaProtheusRodalog(
+                        new ProtheusRodalogRequesterImpl(),
+                        sistemaKey,
+                        integradorProLog,
+                        userToken);
             default:
                 throw new IllegalStateException("Nenhum sistema encontrado com a chave: " + sistemaKey.getKey());
         }