+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Change Log
 ==========
+### Stage
+* Cria integração de ordens de serviço do checklist com o sistema Transport da Translecchi
 
 <a name="v0.0.90"></a>
 ## Version [v0.0.90](https://github.com/luizfp/PrologWebService/compare/v0.0.89...v0.0.90) (release-date) [unreleased]
@@ -14,7 +16,6 @@ Change Log
 ### Refactors
 * Permite atualizar o tipo de um veículo se não houverem pneus aplicados (PL-1903)
 * Verifica permissão de relatórios de quiz na busca dos modelos da unidade
-
 
 <a name="v0.0.88"></a>
 ## Version [v0.0.88](https://github.com/luizfp/PrologWebService/compare/v0.0.87...v0.0.88) (2019-04-11)